refactor: Remove old user model and reorganize user model structure
